This script checks the version of the "Lasernet-Service" and "Lasernet-Client"
The script retrieves the path of the "c:\Program Files\" directory and then searches for all Lasernet folders within the "Formpipe Software" subdirectory. It iterates through each Lasernet folder and retrieves the version information of the specified executables: "LnService.exe" and "LnClient.exe". The script displays the version information for each executable.